reorganize test fixtures and improve login state reset functionality
