+++ app/tests/conftest.py
@@ -8,20 +8,19 @@ from kivy.clock import Clock
 from kivy.config import Config
 from kivy.lang import Builder
 
+from app.screens.utils.utils import call_db, get_sha
+
 Config.set("graphics", "width", "400")
 Config.set("graphics", "height", "600")
 Config.set("graphics", "window_state", "hidden")
 Config.set("kivy", "exit_on_escape", "0")
-
-project_root = Path(__file__).parent.parent
-sys.path.insert(0, str(project_root))
 
 
 @pytest.fixture(scope="session", autouse=True)
 def setup_kivy():
     EventLoop.ensure_window()
 
-    base_path = Path(__file__).parent.parent / "app" / "ui"
+    base_path = Path(__file__).parent.parent / "ui"
     Builder.load_file(str(base_path / "loading.kv"))
 
     yield
@@ -64,3 +63,33 @@ def dump_stacks_on_hang():
     faulthandler.dump_traceback_later(20, repeat=False, file=out)
     yield
     faulthandler.cancel_dump_traceback_later()
+
+
+@pytest.fixture
+def reset_login_state(kivy_app):
+    """Reset login screen to clean state before each test"""
+    sm = kivy_app.root
+
+    if sm.current != "login":
+        sm.current = "login"
+        Clock.tick()
+
+    login_screen = sm.get_screen("login")
+    Clock.tick()
+
+    login_screen.ids.word_input.text = ""
+    login_screen.key = ""
+
+    call_db("DELETE FROM passwords WHERE destination='login_test'")
+
+    test_password = "test_dev_pass_123"
+    test_password_hash = get_sha(test_password)
+    call_db(f"INSERT INTO passwords VALUES ('login_test', '{test_password_hash}')")
+
+    login_screen.passwords = login_screen.db.get_login_password("login_test")
+
+    Clock.tick()
+
+    yield login_screen
+
+    call_db("DELETE FROM passwords WHERE destination='login_test'")
