+++ pytest.ini
@@ -1,5 +1,6 @@
 [pytest]
-testpaths = tests
+pythonpath = .
+testpaths = app/tests
 python_files = test_*.py
 python_classes = Test*
 python_functions = test_*
